+++ client/src/textures/common_texture/dynamic_sprite.cpp
@@ -20,8 +20,8 @@ DynamicSprite::DynamicSprite(std::vector<std::pair<int, int>> spritesCoordTape,
 DynamicSprite::~DynamicSprite(){}
 
 /*
-Devuelve un area (Area) con las coordenadas y dimensiones del proximo a sprite a 
-ser usado.
+Devuelve un area (Area) con las coordenadas y dimensiones del 
+proximo a sprite a ser usado.
 */
 Area DynamicSprite::getNextArea(){
     float xCoord = this->spritesCoordTape[this->index].first;
@@ -42,5 +42,5 @@ sprite de la tira de sprites que los constituyen;
 false en caso contrario.
 */
 bool DynamicSprite::is_last_sprite(){
-    return ((this->index + 1) >= this->spritesCoordTape.size());
+    return ((this->index + 1) >= (int)this->spritesCoordTape.size());
 }
